zfs_acl: split header into core and platform-specific variants
The common set of function names in the same "core" header makes it look
like these are all needed for all platforms, but really only a small
handful of functions are used by core code. Separating them makes that
much easier to see.

Sponsored-by: TrueNAS
