+++ module/os/freebsd/zfs/zfs_acl.c
@@ -44,7 +44,7 @@
 #include <sys/policy.h>
 #include <sys/zfs_znode.h>
 #include <sys/zfs_fuid.h>
-#include <sys/zfs_acl.h>
+#include <sys/zfs_acl_impl.h>
 #include <sys/zfs_dir.h>
 #include <sys/zfs_quota.h>
 #include <sys/zfs_vfsops.h>
