+++ module/os/freebsd/zfs/zfs_dir.c
@@ -48,7 +48,7 @@
 #include <sys/callb.h>
 #include <sys/smp.h>
 #include <sys/zfs_dir.h>
-#include <sys/zfs_acl.h>
+#include <sys/zfs_acl_impl.h>
 #include <sys/fs/zfs.h>
 #include <sys/zap.h>
 #include <sys/dmu.h>
